+++ golang-crosscompile-build.bash
@@ -11,7 +11,7 @@ version=${versiond//./_}
 #Remove White Space
 version=${version// /}
 versiond=${versiond// /}
-platforms="windows/386 windows/amd64 linux/386 linux/amd64 linux/arm darwin/386 darwin/amd64"
+platforms="windows/386 windows/amd64 linux/386 linux/amd64 linux/arm darwin/amd64"
 printf " ---- Building Supportworks Request Import $versiond ---- \n"
 rm -rf "release/"
 rm -rf "builds/"