+++ main.service.go
@@ -7,7 +7,7 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/hornbill/goApiLib"
+	apiLib "github.com/hornbill/goApiLib"
 )
 
 //getCallServiceID takes the Call Record and returns a correct Service ID if one exists on the Instance
@@ -52,13 +52,15 @@ func searchService(serviceName string, espXmlmc *apiLib.XmlmcInstStruct, buffer 
 	espXmlmc.SetParam("entity", "Services")
 	espXmlmc.SetParam("matchScope", "all")
 	espXmlmc.OpenElement("searchFilter")
-	espXmlmc.SetParam("h_servicename", serviceName)
+	espXmlmc.SetParam("column", "h_servicename")
+	espXmlmc.SetParam("value", serviceName)
+	espXmlmc.SetParam("matchType", "exact")
 	espXmlmc.CloseElement("searchFilter")
 	espXmlmc.SetParam("maxResults", "1")
 
-	XMLServiceSearch, xmlmcErr := espXmlmc.Invoke("data", "entityBrowseRecords")
+	XMLServiceSearch, xmlmcErr := espXmlmc.Invoke("data", "entityBrowseRecords2")
 	if xmlmcErr != nil {
-		buffer.WriteString(loggerGen(4, "Unable to Search for Service: "+fmt.Sprintf("%v", xmlmcErr)))
+		buffer.WriteString(loggerGen(4, "Unable to Search for Service: "+xmlmcErr.Error()))
 		//log.Fatal(xmlmcErr)
 		return boolReturn, intReturn
 	}
@@ -66,14 +68,14 @@ func searchService(serviceName string, espXmlmc *apiLib.XmlmcInstStruct, buffer 
 
 	err := xml.Unmarshal([]byte(XMLServiceSearch), &xmlRespon)
 	if err != nil {
-		buffer.WriteString(loggerGen(4, "Unable to Search for Service: "+fmt.Sprintf("%v", err)))
+		buffer.WriteString(loggerGen(4, "Unable to Search for Service: "+err.Error()))
 	} else {
 		if xmlRespon.MethodResult != "ok" {
 			buffer.WriteString(loggerGen(5, "Unable to Search for Service: "+xmlRespon.State.ErrorRet))
 		} else {
 			//-- Check Response
 			if xmlRespon.ServiceName != "" {
-				if strings.ToLower(xmlRespon.ServiceName) == strings.ToLower(serviceName) {
+				if strings.EqualFold(xmlRespon.ServiceName, serviceName) {
 					intReturn = xmlRespon.ServiceID
 					boolReturn = true
 					//-- Add Service to Cache
@@ -85,6 +87,11 @@ func searchService(serviceName string, espXmlmc *apiLib.XmlmcInstStruct, buffer 
 					newServiceForCache.ServiceBPMChange = xmlRespon.BPMChange
 					newServiceForCache.ServiceBPMProblem = xmlRespon.BPMProblem
 					newServiceForCache.ServiceBPMKnownError = xmlRespon.BPMKnownError
+
+					//--- Hacky, but required as the SM Service entity doesn't return Release Request BPM name. Can be removed after the next SM release, July 1st 2021
+					newServiceForCache.ServiceBPMRelease = getReleaseBPM(intReturn, espXmlmc, buffer)
+					//---
+
 					serviceNamedMap := []serviceListStruct{newServiceForCache}
 					mutexServices.Lock()
 					services = append(services, serviceNamedMap...)
@@ -95,4 +102,35 @@ func searchService(serviceName string, espXmlmc *apiLib.XmlmcInstStruct, buffer 
 	}
 	//Return Service ID once cached - we can now use this in the calling function to get all details from cache
 	return boolReturn, intReturn
+}
+
+func getReleaseBPM(serviceID int, espXmlmc *apiLib.XmlmcInstStruct, buffer *bytes.Buffer) (releaseBPMID string) {
+	//-- ESP Query for service
+	espXmlmc.SetParam("application", appServiceManager)
+	espXmlmc.SetParam("queryName", "basicServiceDetails")
+	espXmlmc.OpenElement("queryParams")
+	espXmlmc.SetParam("serviceId", strconv.Itoa(serviceID))
+	espXmlmc.CloseElement("queryParams")
+	espXmlmc.OpenElement("queryOptions")
+	espXmlmc.SetParam("queryType", "logRequestBPM")
+	espXmlmc.CloseElement("queryOptions")
+
+	XMLServiceSearch, xmlmcErr := espXmlmc.Invoke("data", "queryExec")
+	if xmlmcErr != nil {
+		buffer.WriteString(loggerGen(4, "Unable to Search for Release BPM: "+xmlmcErr.Error()))
+		return
+	}
+	var xmlRespon xmlmcServiceListResponse
+
+	err := xml.Unmarshal([]byte(XMLServiceSearch), &xmlRespon)
+	if err != nil {
+		buffer.WriteString(loggerGen(4, "Unable to Search for Release BPM: "+err.Error()))
+		return
+	}
+	if xmlRespon.MethodResult != "ok" {
+		buffer.WriteString(loggerGen(5, "Unable to Search for Release BPM: "+xmlRespon.State.ErrorRet))
+		return
+	}
+	releaseBPMID = xmlRespon.BPMRelease
+	return
 }